# shared-contracts/openapi.yaml
openapi: 3.1.0
info:
  title: GS AI Platform API
  version: 0.2.0
  description: >
    Contract-first API for the GS AI platform. Native clients (Kotlin/Compose,
    Swift/SwiftUI) consume this contract via Ktor. Streaming uses SSE; WebSocket
    upgrade is planned for interruptible/voice sessions.
servers:
  - url: /api
    description: Relative to deployment origin
...
paths:
...
  /v1/uploads:
    post:
      tags: [attachments]
      summary: Upload one attachment (multipart/form-data)
      description: >-
        PHASE 5 input foundation. Stores the bytes and returns a real
        Attachment record. The id must later be claimed by
        POST /v1/conversations/{id}/messages (attachments[]). Reception only —
        no analysis exists at this contract version.
      operationId: uploadAttachment
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                conversationId: { type: string }
                displayName: { type: string }
      responses:
        '201':
          description: Stored
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttachmentUploadResult'
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '413':
          description: File exceeds the 10 MB limit
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '415':
          description: MIME type not in the receive allowlist, or contents do not match the declared type
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '429': { $ref: '#/components/responses/RateLimited' }
...
components:
...
  responses:
    BadRequest:
      description: Malformed or invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    RateLimited:
      description: Client exceeded a rate limit
      headers:
        Retry-After:
          schema: { type: string }
          description: Seconds until the client may retry
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    Attachment:
      type: object
      required: [id, kind, displayName, mimeType, byteSize, createdAt, url]
      properties:
        id: { type: string, format: uuid }
        kind:
          type: string
          enum: [image, pdf, document]
          description: >-
            Extensible: audio | video | screenshot are reserved for when a real
            processor exists. Do not emit them today.
        displayName: { type: string, maxLength: 200 }
        mimeType:
          type: string
          enum: [image/jpeg, image/png, image/webp, image/gif, image/heic, image/heif,
                 application/pdf, text/plain, text/markdown, text/csv]
        byteSize:
          type: integer
          maximum: 10485760
        createdAt: { type: string, format: date-time }
        url:
          type: string
          description: Relative fetch path, e.g. /api/v1/files/:id
    AttachmentUploadResult:
      type: object
      required: [attachment]
      properties:
        attachment: { $ref: '#/components/schemas/Attachment' }
...
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }